Fix the problem of the missing term "(x^34)*(34!/34!)" in maclaurin summation (initially intended to be fixed on a separate branch).
Essentially, we are computing a sum of element 0 to 33.
In other words: "x^0/0! + x^1/1! + ... + x^33/33!".

However, in order to avoid the information-loss embedded in division, we should first compute "x^0*(33!/0!) + x^1*(33!/1!) + ... + x^33*(33!/33!)", and then divide the result by 33!.

This is what we do, but unfortunately, we do it with 34! instead of 33!.
In other words, each factor in the series is unnecessarily multiplied by 34 (including the factor which the intermediate sum is divided by at the end).

So this fix is very simple - reduce each factor by 34.
All outcomes should remain identical.
However, the fact that the factors are now smaller allows for higher maximum values of "x".
More precisely, it allows it for some cases with "precision" between 0 and 5.
But since our minimum precision is 32, there should not be any difference in output whatsoever.